gh-93649: Split float/long tests from _testcapimodule.c
