+++ Modules/_testcapi/parts.h
@@ -33,6 +33,8 @@ int _PyTestCapi_Init_DateTime(PyObject *module);
 int _PyTestCapi_Init_Docstring(PyObject *module);
 int _PyTestCapi_Init_Mem(PyObject *module);
 int _PyTestCapi_Init_Watchers(PyObject *module);
+int _PyTestCapi_Init_Long(PyObject *module);
+int _PyTestCapi_Init_Float(PyObject *module);
 
 #ifdef LIMITED_API_AVAILABLE
 int _PyTestCapi_Init_VectorcallLimited(PyObject *module);
