+++ PCbuild/_testcapi.vcxproj
@@ -104,6 +104,8 @@
     <ClCompile Include="..\Modules\_testcapi\docstring.c" />
     <ClCompile Include="..\Modules\_testcapi\mem.c" />
     <ClCompile Include="..\Modules\_testcapi\watchers.c" />
+    <ClCompile Include="..\Modules\_testcapi\float.c" />
+    <ClCompile Include="..\Modules\_testcapi\long.c" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc" />
